refactor: Fetch객체 및 메서드 네이밍 수정, 생성자 구현
