+++ BoxOffice/Controller/ViewController.swift
@@ -4,8 +4,8 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        Fetcher().fetchBoxOffice(targetDate: "20240219")
-        Fetcher().fetchMovieInfo(movieCode: "20231415")
+        MovieAPIFetcher().fetchBoxOffice(targetDate: "20240219")
+        MovieAPIFetcher().fetchdetailMovieInfo(movieCode: "20231415")
         checkJSONLoader()
     }
